fixes  availability zone filter when creating a share from snapshot
Fixes bug #1915706 that was discovered in  Manila UI.
Issue is invalid input error being raised, even when
availability_zone provided is the same, as parent share's.
The reason is availability_zone is name and source_share_az
is id in the filter [1]. Now, regardless of availability_zone
provided is id or name, it will be always converted to a name.
And if availability_zone is the same as parent share's
operation will complete successfully.

[1] https://github.com/openstack/manila/blob/2e27c71877fdb601ba90b44c3573d8e50e099042/manila/share/api.py#L200

Closes-Bug: #1915706
Change-Id: I5039b7adf563d926053b423a4cc6e175d3f1e5cf