[ci] Temporarily set docs job to non-voting
We're seeing sporadic failures of this job and
the rate of failures is increasing in the past
couple of days. Since we're at feature freeze and
the doc is being built fine, we decided to set
this job to non-voting and investigate the failure
while allowing feature patches to merge.

Partial-Bug: #1918707
Change-Id: I2233f951f87e38d82d7963a57695e9cf84cbf558
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>