Filter shares by share type "extra_specs"
This change fixes the query being used to filter share
types. We were using an "or" clause instead of an "and" clause,
and that was causing all the extra specs to be shown instead of
the ones that the user specified. With this fix, manila users will
be able to filter shares by share type extra_specs accurately.

Closes-Bug: #1929121

Change-Id: I5cf9788dcafebdf88426617d2824f4be42dbb412
(cherry picked from commit 4fbacf1df1dbe7a85726e22892d462b5b7c4e316)