Deprecate create snapshot from share group in CephFS
Deprecate the create snapshot from share group feature
in CephFS drivers (both Native and NFS Ganesha).

This needs to be done since the subvolume group snapshot
feature is no longer supported in mainline CephFS, meaning that,
if we leave the call in place, an exception will be raised
from the backend and lead to an error status for the attempted
snapshot from share group.

Existing group snapshots can still be listed and deleted, so these
operations stay the same.

Closes-Bug: #1925332
Change-Id: I49efb7042d38257416875f40c03923e974b2eb8e