[NetApp] Fix list of mandatory services for CIFS share creation
To create a CIFS share, it is mandatory to configure an active
directory service, but the code was requiring to configure also
the `ldap` and `kerberos` services to be configured during the
creation of a CIFS share.

This patch fixes this, removing `ldap` and `kerberos` from the
list of mandatory security services.

Change-Id: I34020da6a929f5c7ae1c63bc0f2ef338c49dbe10
Closes-bug: #1942124