Fix ignored [neutron] options
The following three options have had no effect since implementation
to use these parameters were unintentionally removed by [1].
 - url
 - url_timeout
 - auth_strategy

This fixes/deprecates these parameters. The details are as follows.

1) url
This option is fixed and now overrides the endpoint url which Manila
uses to connect neutron. Default of the url option is changed to keep
the existing behavior by default.

2) url_timeout
This option is just deprecated, because there is already the timeout
parameter in keystoneauth to define request timeout.

3) auth_strategy
This option is also deprecated. The auth mechanism should be defined
by the auth_type parameter in keystoneauth instead.

[1] 7fc492ea79ab0dbbc3d50bf477ae2a1586281dda

Closes-Bug: #1946990
Change-Id: I34cfff78d70a5e1cbd2ee75da36b692cb19a45a4