Adapt CephFS driver to do not try to escape export ip
Changes CephFS driver to do not try to escape export ip in case it
is a hostname. The CephFS driver might use the hostname as an
export ip in case no ``cephfs_ganesha_server_ip`` is configured.
If that happens, we should not be allowing ip validations to occur
uppon a hostname, because it differs from an IPv4 or IPv6 and that
isn't necessarily an issue.

Change-Id: I1bd57eec3474768596cfab1bf2f8eb6123cc95a8