Force disk wipe when running lvcreate
The LVM driver may fail while creating new volumes using the
lvcreate command. Sometimes it can say an existing ext4 filesystem
was found in the destination, but we shouldn't be worried about it.
This change modifies the behavior and forces the lvcreate command
to wipe the space if a signature was found.

Change-Id: I4c10bcdfc499452d4b05dfbee8059be35dfe5586
(cherry picked from commit 89028725d1ac6890332f4c06a30c2416ea951aff)