Modify docker instalation for fedora systems
Use install_package instead of using a specific package manager
while installing docker on fedora systems. With DNF also a
possibility, the better option is to do not take a stand in chosing
which package manager will be used to install packages.

We have a install_package function that will use the appropriate
package manager to install the programs we need to install.

Change-Id: I0524554dae2d6970191ba3afa583480c391b5c3c