Container: Multiple subnets per AZ
This change gives the Container driver the ability to create share
servers that span multiple subnets in the same availability zone, where
previously it was only possible to do so with one subnet per
availability zone. It also enables the driver to create new subnets on
share networks that already have shares exported on them.

Some relevant changes in the driver behavior:

- container creation is now separate from its start
- containers no longer use the default Docker network; instead,
  dedicated networks are created (one for each subnet)
- fetch_container_addresses now returns a list of addresses
- find_container_veth, _get_veth_state, _get_corresponding_veth were
  removed
- these new methods were written:

  - start_container
  - fetch_container_interfaces
  - create_network
  - remove_network
  - connect_network
  - disconnect_network
  - get_container_networks
  - get_container_veths
  - get_network_bridge
  - get_veth_from_bridge
  - _setup_server_network

DocImpact
Depends-On: I7de9de4ae509182e9494bba604979cce03acceec
Partially-Implements: blueprint multiple-share-network-subnets
Signed-off-by: Eduardo Santos <eduardo.experimental@gmail.com>
Change-Id: I81a616b57c95508f30d872383f5c2d68718950d7