Add grenade-skip-level job
This adds a grenade job that specifically skips the last release,
to prove that we can do it. Such a job would facilitate a new
goal of a tick/tock release arrangement where slower-moving operators
could stick on the "tick" release and know that they can upgrade to
the next "tick" release directly.

See https://review.opendev.org/c/openstack/governance/+/828777
for a proposal on why this is needed, and
https://review.opendev.org/c/openstack/grenade/+/826101
for a similar skip level job that's being created from the
integrated grenade job

Change-Id: I8c1584176643a8086ffb23358176073a18c82b45
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>
Depends-On: If4820dd4ec7e05ade6eb4a82a328797102704570