Add missing deprecated messages to policy
During the Yoga cycle, we've introduced a new API to check if it
is possible to add share network subnets to share networks that
are currently being used. We missed the addition of the deprecation
messages to that new policy though. This change serves as a follow
up to add these missing items.

Change-Id: I163e2691105468a0e821c8a5991f8c20e85ff599