Fix GaneshaNASHelper update_access signature
The gluster GaneshaNFSHelper update_access method calls
into the GaneshaNASHelper update_access method, but the
signature it uses for this call is not up to date with
the 2017 dynamic export changes made to the Ganesha
driver.

Fix this by adding an update_access method to
the gluster-specific GaneshaNFSHelper which accepts
the signature that the Gluster driver expects but calls
into Ganesha with the signature that it expects.

[1] Change-Id: I01ec100c0afe28a84e9afa8e0660d299e4b3d160

Closes-Bug: #1964696
Change-Id: I2cae5ba54ae0a95e84d5ef7cf0f05287b7ae5d22
(cherry picked from commit 6f4191a6d92342c35ece64c735be6c43ca171d58)