NetApp: Disable tunneling in vserver_exists
When share server is in error state in Manila and backend vserver has
not been created, there exist entries of share server and shares in
database. Deletion of share in such case cause share in error_deleting
state. In order to delete share correctly, we need to fix _get_vserver
which should check vserver existence before building the vserver
client.

Closes-Bug: #1964592
Change-Id: I1f6b94a7e54403a559a5d77b4d93b939a40e76bf