[CI] Install dependencies for docs target
When tox 'docs' target is called, first it installs the dependencies
(listed in 'deps') in 'installdeps' phase, then it installs manila
(with its requirements) in 'develop-inst' phase. In the latter case
'deps' is not used so that the constraints defined in 'deps' are not
used. This could lead to failures on stable branches when new packages
are released that break the build. To avoid this, the simplest solution
is to pre-install requirements, i.e. add requirements.txt to 'docs' tox
target.

Conflicts:
  tox.ini

NOTE(elod.illes): conflict is due to branch specific upper constraints
file link.

Change-Id: I241ab2788b870060e1956a8fc029705378a2a8f4
(cherry picked from commit 9b4c8c92518c25e985c9649b0a7cb430169db712)