Simplify composition of processutils.execute kwarg
Currently we ensures the root_helper key when the run_as_root key is
included, but the logic does not check whether run_as_root has a True
value.

So even if the function is called with run_as_root=False, the logic
ensures root_helper, and this is just redundant.

Because root_helper has no effect unless root_as_root is defined and
has a True value, we can always pass root_helper.

Change-Id: Ic78d11d36dc2d0b52176a5ac2c49cb06d59476db