Update image URL and image name
In order to switch our manila service image to Ubuntu Jammy, we
needed to get in a couple of changes that Modified the flavor for
the manila service image. After all changes got merged, we switched
to the Ubuntu Jammy image [1].

However, now the manila service image master on Tarballs is using
a Ubuntu Jammy image, which requires more resources. It would not
be ideal to backport the changes that modify the flavor to the
stable branches because it is not an actual requirement for them,
and that also forced us to reduce concurrency of some jobs on the
master branch.

Using the Ubuntu Jammy image broke our stable branches CI, since
with less resources we are not able to successfuly spawn a nova
instance using the Ubuntu Jammy image.

In order to avoid that, this change will modify the manila service
image to be used in the stable branches. This selected version
is the latest built whilst still using Ubuntu Focal. This will
allow tests to pass on stable branches, since Manila will be able
to spawn VMs using this image.

[1] https://review.opendev.org/c/openstack/manila-image-elements/+/841264

Change-Id: Ibb1b7ad8b167c469e213c5de0ce56214c1fcff2b