Fix manila plugin.sh set backend_availability_zone
Manila plugin is setting the availability zone of the
backends without checking if it is set. As result, a CI
configuring that field will not be respected.

Fixed by only setting the field in case its value was
not set before. Otherwise, keep as it is.

Change-Id: I546e5f057f3b748417dcbcfe9c3f3ae4a5d51556
(cherry picked from commit 79af79f07b66b061ca1c47e6c5055eeaff5bd454)