Fix stable/train CI jobs
This is a combination of two cherry-picks
to ensure CI runs and does so with constraints
that are unfortunately required on this
branch.

1) Define queues at project level

Defining queues at project.pipeline level was deprecated in release
4.1.0 of Zuul [1] and support for it will (probably) be dropped in
release 7.0.0. [2][3]

This change is a stable/train only patch since newer branches use
layout.yaml inside zuul.d dir to declare project check/gate jobs.
See topic [4] for more info related to this change.

[1] https://zuul-ci.org/docs/zuul/latest/releasenotes.html#relnotes-4-1-0-deprecation-notes
[2] https://lists.openstack.org/pipermail/openstack-discuss/2022-May/028603.html
[3] https://lists.zuul-ci.org/pipermail/zuul-discuss/2022-May/001801.html
[4] https://review.opendev.org/q/topic:fix-queue-config

2) Pin tempest plugin and scenario test image

Original Change-Id: Ic7213030dd433b765d5a815d8a31aee01b09da1c

Tempest is pinned to a 26.1.0 on this stable
branch [5]. manila-tempest-plugin in Zed requires a newer
tempest release [6] than this one. So, pin
manila-tempest-plugin as well to a version that can still
work with the pinned version of tempest.

Also, manila's test image (built from manila-image-elements)
is based off of Ubuntu 22.04 (Jammy Jellyfish) in the Zed
release; we don't need that image to be used to test this
stable branch and we could use an older version (based on
Ubuntu 20.04 - Focal Fossa) that requires less memory/CPU [7]

[5] https://review.opendev.org/c/openstack/devstack/+/838051
[6] https://review.opendev.org/c/openstack/manila-tempest-plugin/+/849995
[7] https://review.opendev.org/c/openstack/manila-tempest-plugin/+/846061

Co-Authored-By: Eduardo Santos <eduardo.experimental@gmail.com>
Change-Id: I0e75dc2285d26c45f2811d7050976b7474863a77
Signed-off-by: Douglas Viroel <dviroel@redhat.com>