Fix InvalidInput wrong message
When performing some operations, there is a chance that the API
will raise a InvalidInput exception. However, sometimes we are
treating this exception alongside others, and expecting these
exceptions to inherit from the same base. The exceptions could
have different attributes to store the message though.

This change modifies the message treatment to stringify the
exception instead of trying to use specific attributes.

Change-Id: I1f54f6be46c5a1e997c8bf589503a5e62a71d644