[CI] Pin tempest plugin and scenario test image
Tempest is pinned to a 26.1.0 on this stable
branch [1]. manila-tempest-plugin in Zed requires a newer
tempest release [2] than this one. So, pin
manila-tempest-plugin as well to a version that can still
work with the pinned version of tempest.

Also, manila's test image (built from manila-image-elements)
is based off of Ubuntu 22.04 (Jammy Jellyfish) in the Zed
release; we don't need that image to be used to test this
stable branch and we could use an older version (based on
Ubuntu 20.04 - Focal Fossa) that requires less memory/CPU [3]

[1] https://review.opendev.org/c/openstack/devstack/+/838051
[2] https://review.opendev.org/c/openstack/manila-tempest-plugin/+/849995
[3] https://review.opendev.org/c/openstack/manila-tempest-plugin/+/846061

Change-Id: Ic7213030dd433b765d5a815d8a31aee01b09da1c
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>