[CI][stable/zed] pin test plugin in grenade skip-level
The version being installed (stable/wallaby) has a
pinned version of tempest in its requirements [1]. This
disallows using the latest manila-tempest-plugin with
this branch. So lets pin manila-tempest-plugin as well
to a version that works with tempest===27.0.0.

This pinning shouldn't matter for our testing
in this job since newer versions of the test plugin
are expected to test only new features for the most
part.

This change is appropriate for stable/zed; since we're
moving main/2023.1/antelope to start with the Yoga release [2]


[1] https://opendev.org/openstack/requirements/src/branch/stable/wallaby/upper-constraints.txt#L393
[2] https://review.opendev.org/c/openstack/manila/+/859875

Change-Id: If6dcfefb3702c1f58969dd5c88e67e967608679e
Signed-off-by: Goutham Pacha Ravi <gouthampravi@gmail.com>