Shorten snapshots names in CephFS drivers
The name set to snapshots in the CephFS drivers
is generated appending the snapshot id with the
snapshot instance id.

This is a problem because Ceph has a limitation
in the name of the subvolumes (shares) being created.

What happens then is, when trying to create a snapshot,
the snapshot gets created with a truncated name and it cannot
be accessed later.

This change drops the snapshot instance id from the
assigned name in order to avoid this limitation.

Closes-Bug: #1967760

Change-Id: Ifab3ec9de2abeb1af8e6499feca0236e1ee22fcf