Raise an exc if no VIP/backend is available
When deploying Manila CephFS NFS with cephadm,
the manila share service fails to start with the error
"Backend cephfsnfs supports neither IPv4 nor IPv6".

This happens because the NFS Ganesha daemon fails
to start for some reason and therefore the driver never
gets the location of the NFS Ganesha service
that will be used as the backend.

We should raise an error in this case with more information
about the failure.

This change adds this to the NFSClusterProtocol helper.

Closes-Bug: #1990150

Change-Id: Ic74c6c9d9770974848ec9dfb6d2225ce9b3bcb52