Rename 'nfs_cluster_id' to 'cluster_id' in the NFSProtocolHelper
The CephFS NFS driver, specifically the NFSProtocolHelper
implementation, is passing a wrong param to the Ceph backend
and this is preventing users to add and/or deny access to the
created shares. The 'nfs_cluster_id' field has been renamed
to 'cluster_id' to comply with what the rados client expects.
With this fix, users of the CephFS NFS NFSProtocolHelper can
normally create and remove access to their shares.

Closes-Bug: #1991378
Change-Id: Ibc5681e59691d55536c5539feef97b14f74d98ba