Add filesystem info to FSAL in CephFS NFS
The filesystem information was missing from the exports
in the CephFS NFS driver. Because of this, the exports
were created on the first available filesystem.

This generates inconsistencies in deployments in which more than
one filesystem is being used.

This patch set adds the filesystem information in the
created exports.

Closes-Bug: #1991938
Change-Id: I639640f738571c77ea9bb1fbf60749c00ffcbd3b