[NetApp] Configure VLAN/MTU through subnet metadata
The NetApp driver can setup a share server VLAN segmentation
and MTU based on the subnet metadata. If the subnet metadata has
`set_vlan`, the driver will configure the segmentation
following the metadata fields. If the subnet metadata has the
`set_mtu` field, the MTU is configured according its value.

If share network subnet does not have those metadata fields,
the segmentation configuration keeps working as before, that's it,
it is set according to the values provided by Manila network plugin.

Change-Id: I69731b187527d8d443893dcc94c604688d73e696