[Cephfs] Fix erroneous share mode override on ensure_shares
In order to ensure that a share existed, we used to send the same
request to Ceph as we would do with the share creation. In that
request, we will set the share mode, which is a value set by
the deployer through the `cephfs_volume_mode` config option.

If the share was already created, Ceph would not override the value
we were forwarding for the permission mode. After a bugfix [1], the
value of mode started being reinforced when sent, even if the
subvolume was already created. If the value for
`cephfs_volume_mode` changed after the share was created, the mode
is now being overwritten after manila startup on ensure_shares.

Fixed this issue by modifying the ensure_shares to only confirm if
the export exists.

[1] https://tracker.ceph.com/issues/54375

Closes-Bug: #2002394
Change-Id: Ic5b5b42b882ce1346b4b46d1b29aa31740933e0e
(cherry picked from commit c61c1977c25e151fd2839febd08a7fdac6b1f7b8)
(cherry picked from commit f8d0e0cad271baaac0f86aa3fe156272deabb48e)