Prevent failure on get quiesce_wait_time
quiesce_wait_time was an option introduced in API version 2.75. It
modified the promote API. In the current implementation, there is
a chance that we will receive a {'promote': null} request body. In
that case, and the quiesce_wait_time change was only accounting
that we would receive a dictionary.

This change prevents such failure to happen by ensuring we are
trying to get quiesce_wait_time from a dictionary.

Change-Id: I9444162a7d4e861e5f11f4d798014c7b90216952