SAPCC: fix postpone "clone split" to last step (#117)
When creating a share from Snapshot and clone split needed, we want to
postpone it to the last step. But it didn't happen as expected. The
split controlling variable is overriden by share type specs. This PR
fixes the issue by explictly reading the corresponding specs option into
the controlling vairable.