Include share_type information in notifications
For the purposes of billing, it's useful to provide the information
about share type in oslo.messaging notifications.

Different share types can have different quality of service, limits,
quotas or even be backed by entirely different backends, so it makes
sense to bill shares of different types differently. For example,
a customer may want to choose to create a share with higher performace
via an appropriate share type and thus be charged more for such a share.

Change-Id: I4b8e9344c36aa874b95fce3e0989d34b690159d6
Signed-off-by: Mariusz Adamski <mariusz.adamski@ovhcloud.com>